requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
lxml==4.9.3
cloudinary==1.36.0
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1